6. change UpdatePost to RestController
