+++ src/test/java/com/together/buytogether/post/feature/UpdatePostTest.java
@@ -1,6 +1,5 @@
 package com.together.buytogether.post.feature;
 
-import com.together.buytogether.member.domain.MemberRepository;
 import com.together.buytogether.post.domain.Post;
 import com.together.buytogether.post.domain.PostFixture;
 import com.together.buytogether.post.domain.PostRepository;
@@ -8,7 +7,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.springframework.util.Assert;
 
 import java.time.LocalDateTime;
 
@@ -45,33 +43,4 @@ public class UpdatePostTest {
 
     }
 
-    private class UpdatePost {
-        MemberRepository memberRepository;
-        PostRepository postRepository;
-
-        public UpdatePost(PostRepository postRepository) {
-            this.postRepository = postRepository;
-        }
-
-        public void request(Long memberId, Long postId, Request request) {
-            Post post = postRepository.getByPostId(postId);
-            post.checkOwner(memberId);
-            post.update(
-                    request.title,
-                    request.content,
-                    request.expiredAt
-            );
-        }
-
-        public record Request(
-                String title,
-                String content,
-                LocalDateTime expiredAt) {
-            public Request {
-                Assert.hasText(title, "글 제목은 필수입니다");
-                Assert.hasText(content, "글 내용은 필수입니다");
-                Assert.notNull(expiredAt, "글 만료일은 필수입니다");
-            }
-        }
-    }
 }
